break birthdate constants into helper function
