+++ cypress/automated-tests/youth-pass/address-section-required-fields.spec.js
@@ -1,17 +1,12 @@
-const faker = require('faker');
+import { getRandomApplicantAge } from '../../common/birthdate-constants';
 
 describe('Youth Pass Address Section Required Fields', () => {
+    const faker = require('faker');
+    
     it('proceeds through an application', () => {
         const applicantZipCode = '02114';
         const youthPassUrl = Cypress.env('youth_pass_url');
-        const todaysDate = new Date();
-        const twelveYearsAgo = todaysDate.getFullYear() - 12; 
-        const eighteenYearsAgo = todaysDate.getFullYear() - 18; 
-        const randomBirthdate = faker.date.between(`${eighteenYearsAgo}-11-02`, `${twelveYearsAgo}-11-01`);
-        const applicantBirthdate = 
-            `${randomBirthdate.getMonth() + 1}/
-            ${randomBirthdate.getDate()}/
-            ${randomBirthdate.getFullYear()}`.replace(/\s+/g, "");    
+        const applicantBirthdate = getRandomApplicantAge().applicantBirthdate12to17;
         const applicantSchoolName = `Automation Testing ${faker.datatype.number()} School`;
         const applicantFirstName = faker.name.firstName();
         const applicantLastName = faker.name.lastName();
